Rename mod to weekendproject/todo
